+++ src/main/java/com/mageddo/dnsproxyserver/docker/Entry.java
@@ -1,0 +1,17 @@
+package com.mageddo.dnsproxyserver.docker;
+
+import com.mageddo.net.IP;
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public class Entry {
+
+  private boolean hostnameMatched;
+  private IP ip;
+
+  public String getIpText() {
+    return this.ip != null ? this.ip.toText() : null;
+  }
+}