+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/SolverLocalDB.java
@@ -4,7 +4,6 @@ import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.Config.Entry.Type;
 import com.mageddo.dnsproxyserver.config.ConfigDAO;
 import com.mageddo.dnsproxyserver.server.dns.Messages;
-import com.mageddo.net.IP;
 import dagger.Lazy;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -35,10 +34,11 @@ public class SolverLocalDB implements Solver {
     }
 
     final var askedHost = Messages.findQuestionHostname(query);
-    final var res = HostnameMatcher.match(askedHost, IP.Version.IPV4, hostname -> {
+    final var questionType = Messages.findQuestionType(query);
+    final var res = HostnameMatcher.match(askedHost, questionType.toVersion(), hostname -> {
       stopWatch.split();
-      final var entry = this.findEntryTo(hostname);
-      if (entry == null) {
+      final var found = this.findEntryTo(hostname);
+      if (found == null) {
         log.trace(
           "status=partialNotFound, askedHost={}, time={}",
           askedHost, stopWatch.getTime() - stopWatch.getSplitTime()
@@ -47,16 +47,17 @@ public class SolverLocalDB implements Solver {
       }
       log.trace(
         "status=found, type={}, askedHost={}, time={}, totalTime={}",
-        entry.getType(), askedHost, stopWatch.getTime() - stopWatch.getSplitTime(), stopWatch.getTime()
+        found.getType(), askedHost, stopWatch.getTime() - stopWatch.getSplitTime(), stopWatch.getTime()
       );
 
-      if (Type.is(entry.getType(), Type.A, Type.AAAA)) {
+      if (found.getType().isAddressSolving()) {
+        final var ip = found.getType().equals(questionType) ? found.getIp() : null;
         return Response.of(
-          Messages.aAnswer(query, entry.getIp(), entry.getTtl()),
-          Duration.ofSeconds(entry.getTtl())
+          Messages.answer(query, ip, questionType.toVersion(), found.getTtl()),
+          Duration.ofSeconds(found.getTtl())
         );
       }
-      return this.solverDelegate.get().solve(query, entry);
+      return this.solverDelegate.get().solve(query, found);
     });
     if (res != null) {
       return res;