+++ src/test/java/com/mageddo/dnsproxyserver/templates/HostnameTemplates.java
@@ -5,4 +5,5 @@ public class HostnameTemplates {
   public static final String ACME_HOSTNAME = "acme.com";
   public static final String COM_WILDCARD = ".com";
   public static final String NGINX_COM_BR = "nginx.com.br";
+  public static final String HOST_DOCKER = "host.docker";
 }