+++ RELEASE-NOTES.md
@@ -1,5 +1,5 @@
 ## 3.15
-* Support for AAAA records on docker container and system solving.
+* Support for AAAA records on Docker container, LocalDB and System solving.
 * Formalize Recursion Available on query responses as DPS supports this feature already, see #392
 * Defined a new process of releasing the stable versions, see "Getting Started -> Release Version Control" 
 for more details